separate appium driver to core
