+++ lib/appium_lib/core/core.rb
@@ -6,7 +6,24 @@ require_relative 'search_context'
 require_relative 'wait'
 require_relative 'device'
 require_relative 'capabilities'
+require_relative 'driver'
 
 require_relative 'common/log'
 
 require_relative 'element/window'
+
+module Appium
+  # convert all keys (including nested) to symbols
+  #
+  # based on deep_symbolize_keys & deep_transform_keys from rails
+  # https://github.com/rails/docrails/blob/a3b1105ada3da64acfa3843b164b14b734456a50/activesupport/lib/active_support/core_ext/hash/keys.rb#L84
+  def self.symbolize_keys(hash)
+    raise 'symbolize_keys requires a hash' unless hash.is_a? Hash
+    result = {}
+    hash.each do |key, value|
+      key = key.to_sym rescue key # rubocop:disable Style/RescueModifier
+      result[key] = value.is_a?(Hash) ? symbolize_keys(value) : value
+    end
+    result
+  end
+end
